Pin espree `=3.4.0` to avoid throw on trailing rest comma.
The comma which is causing a problem here is a trailing comma after
a rest, within a function argument's parameter destructuring:

		function a({
		  a = false,
		  ...b,
		}) {
		  // ...
		}

This was previously allowed with `experimentalObjectRestSpread` enabled
but now throws an error with https://github.com/eslint/espree/issues/310.

I'll be removing the comma as well as it's clearly a point of contention
but this will allow the old docs to get deployed.